Refinements to form setup, adding support for popup save
